Changing the deployment folder destination name
When custom templates provided during the overcloud deploy,
the name of the copied folder should not be changed
because if the overcloud_deploy script provided as well, it has
hardcoded path to the network-environment.yaml file, so the change
of the template folder name will fail the proper execution of the
overcloud_deploy script.

The 'set overcloud_deploy script' task should be executed if the
overcloud_deploy script provided as well, otherwise the path to
the overcloud_deploy script will not be correct.

Change-Id: I3a72724ecd6938875df8708ccb8c6f0be32b3bba