Fix undefined variable in rdo-release
Replaced product_version undefined variable with the install.version.

Additionaly fixed image modification in virsh provisioner to
allow cloud-user to become any user (not only root). This is needed
to run commands as the 'stack' user in the underclodu plugin.
This issue is actual when centos image is used for VM's.

RHOSINFRA-940

Change-Id: I77d1daaff2323c27f080783860e17a1f956ea6d1