Restore fact saving for storage
This patch restores fact saving to the files for the
storage node.

Storage templates requires information about nodes disks
We can get that information from virsh, but in case of baremetal
nodes we won't be able to collect facts from these nodes, because
these nodes will be shutdown due to the introspection.

Need to revisit storage templating to avoid facts saving for storage
nodes.

Change-Id: Id3d72922ebd09528bea859da74767e02766ef7a6