Reverts uses of '--os-cloud' with 'source stackrc'
Temporary reverted some of the changes that made by commit
b177f90143079e5054caa4fb2e69040257a4b86b (RHOSINFRA-635).
Restored the use in 'source stackrc' instead of passing a cloud file
into '--os-cloud' parameter.
The reason for the revert is that the use of '--os-cloud' is not
supported in OSP version <= 7.
The creation of the cloud file with all data stays.

Change-Id: I47643e54f0b64fdc0e916dbb2023bd455880727a