Container images environment generate
The DockerNamespace argument is not respected anymore,
generating the current image list file by the
`openstack overcloud container image prepare`
which will point to the undercloud registry.

We always expected to use the undercloud registry after this change,
but we will use the mirrors to populate the undercloud registry.

See Also:
  http://lists.openstack.org/pipermail/openstack-dev/2017-July/119880.html

Change-Id: I2760c6baa782794a629af83a9f22b9191717eb78