Tagging for hybrid mode
Currently infrared determines what flavor to use for a spcific roles by
the group presence in inventory. For example, if compute group is in
inventory then the 'compute' flavor will be used for the nodes.
But for hybrid mode the baremetal host are not getting added to the
inventory initially.

Tagging is reworked to make sure we have flavors only for the named
hosts defined in the instackenv file. Infrared then check if we have
flavor that match to the host and use it even if host is not in ansible
inventory. If flavor is not present then infrared uses
the baremetal flavor.

Change-Id: Ib4d2a4bb36ae58e4a5adea516653a13db9fd640f