Modify external-vlan assignment to prevent packet drops to external api
Currently, external-vlan argument sets the external
default gateway as ip address with 32 mask, in order to reach the
external api ip in case undercloud unable to reach it.
This leads to continuos network lags and packet drops on the undercloud.
Changing it, to manually provide specific ip address from the
external api network range.
As explaining in the plugin.spec, the used address should be excluded
from the external api network pool in order to prevent network conflict.

Change-Id: I238796622afdb6fd295a67d9406ddfc7f692756d