Set custom ceph pg_num and pgp_num
BZ#1481821 introduced 128 as the default value for ceph pg_num
and pgp_num but this is not suitable for the testing enviroments
that we're using(3 ceph osd nodes with 1 osd/node). This change
sets the pg_num and pgp_num to 32 so the ceph cluster reports a
HEALTH_OK status for testing environments.

Change-Id: Iab5b6ec23e952918ddbc998433ced52001fd9115