ext-lb: Do not define resource_registry in external-lb.yaml template
The section resource_registry is defined for OSP in:
openstack-tripleo-heat-templates/environments/external-loadbalancer-vip.yaml,
with correct values. The current values defined in external-lb.yaml.j2 cause
that instances do not get Fixed IP dhcp lease. I am not sure why we defined
this section for OSP11 but probably for something which was fixed meanwhile.

Change-Id: I141ed6999733727ea8dfe4f641123e6f908a8090