Add port on storage network to the monitor composable role
This change adds the port for the storage network in the resource
registry for the monitor composable role. In addition it removes
the storage management network from the nic template as it should
not be required for the monitor nodes per:
http://docs.ceph.com/docs/master/rados/configuration/network-config-ref/
It also sets dhcp to false for the 3rd nic of the monitor nodes to
avoid multiple default routes.

Change-Id: Ib0d71999a48ec0fbdc2ccbd51ad0bf81c2dc1473