Physnet type for hybrid environment
The tripleo-overcloud plugin creates an external network.
For virt env it uses flat physnet type and for baremetal uses vlan.
The decision of which type to use is based on the inventory groups.
In case 'hypervisor' or 'bmc' group exist in inventory, flat is
used. Otherwise means it is baremetal env and vlan type is used.

Hybrid environment has both virtual and baremetal nodes,
so 'hypervisor' group exists but vlan type should be used.
All the network in hybrid exposed so vlan type should be used.
Adding condition in which flat type will be used if the virt
deployment is not hybrid.

Change-Id: Iba91919546a62b9074a55bacd26500dd8a59d885