Fix add_overcloud_host
Sometimes the hypervisor_hostname value from the facts can be null (None).
In that case calling of the |trim filter on it converts that value to non-empty
'None' and thus our current check passes and allows to use incorrect value.

Added additional check to verify that hostname is not None and only then
verify that it is not empty.

RHOSINFRA-1747
BUGFIX

Change-Id: If895dc802610f908b3399307dda87964bca7308b