avoid ignore_errors on check subscription
Replace use of ignore_errors with failed_when
on check subscriptions tasks because error
return code is expected result and not a real error.

This does not change the cdn_status.rc value or the
ansible logic but it does change the output.

Avoids use of success filter because task will always be succesful now. Instead use .rc result which contains right information.

Now ansible reports this tasks as succeded (green)
instead of failed and ignored (red).

The purpose of the task is to check the subscription status, not
if the result code from the command execution is zero.

Change-Id: I66b19adeb12c86f60dc0301ea706b8439be847a1
Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>