Update unused groups after scale down
After scale down operation the unused node was
completely removed from the inventory file.

This is not quite right because we still can re-use that node.
Additionally this node is required to be present if we want to
do a scale-up operation and add another node to the inventory.
Old node in that case is required to calculate a new node start index.

This patch makes sure that the scaled-down nodes is saved to the unused
group and remains in the inventory file.

RHOSINFRA-1927

Change-Id: I512f7a94b9ed92cc8004251b45e006aaf6ae8436