Enhance undercloud reboot procedure.
Rebooting bm undercloud takes longer than vm so procedure has
to be adjusted.

Current mechanism of asserting if system is up is not reliable.
This patch changes it to call wait_for module in do..untill loop:
 try to connect to server to get 'OpenSSH' string with a 30 seconds
timeout, retrying this for 21 time. This results in
 30x21 = 630 sec for whole operation. This might look too high
but better to wait 10 minutes rather than re-run reprovisioning.

But before start check for port being accessible - ensure that node
 went down by not replying to ICMP ping.
That's needed cause it takes significant amount of time to shutdown
all OpenStack services.

Values for virtual environments left intact.

Change-Id: I033bf4a142bad46e729481177f9450bda1d13dda