Fix condition for UC packages installation.
The tripleo-overcloud plugin allows installing custom packages
into the undercloud node (with --undercloud-packages flag).

This patch fixes condition that checks if those packages
should be installed into the undercloud machine itself or
into the containers. For OSP13 and below (including OSP13)
we need to install packages to the machine.

RHOSINFRA-2126

Change-Id: Ibd0c983557df2948fdc9239c2cf886e9438f1a43