Override registry params from CLI
Currently, when we deploy OSP we read container registry
paramters from a file in puddle repo. The problem is that we
cannot override these parameters from the CLI.

This patch fixes that problem. It replaces default
parameters from the puddle file with the options
provided by the user with the CLI arguments (--registry-*)

This allows to override, for example, ceph related parameters like
image, tag and namespace.

RHOSINFRA-2126

Change-Id: I6360f9a379ae49271e9abfbdbdc85d1e1d07ee90