Process undercloud.conf before retrieving undercloud_public_host
The task that overrides values in undercloud.conf may very well
be modifying network values such as undercloud_public_host
to refer to a different network.  This has to happen before
the ssl.yml reads the file in order to estalbish the hostname
that should be copied into the SSL certificate.

Change-Id: I1f34c5e181e166f3254ff61ee560d71c0e6602c7
Resolves: https://github.com/redhat-openstack/infrared/issues/336