Update manila environment file names
Containerized manila is now the standard way to deploy
so we can avoid confusion by just using the regular
manila environment file names for these rather than having
some manila environment files with 'docker' in their names.

Note that we are using the content of
environments/manila-cephfsnative-config-docker.yaml
for
environments/manila-cephfsnative-config.yaml.  This resolves an
issue wherein the parameters in the latter file were not updated
when support for the CephFS-NFS back end was added [1].

Closes-Bug: #1782622

[1] Ib408c7827e5fba0c1b01388db26363806fc64370

Change-Id: I9197d3980468d0835c18260ba4529b8e158eb66f