Only wait for UC to go down when we're really rebooting
The reboot task is guarded by "update is changed" condition, but
waiting for the node to go down isn't guarded this way, so we used to
wait for undercloud to go down even when it wasn't rebooting. This is
now fixed by adding the "update is changed" condition to the waiting
task too.

Change-Id: I65d1bd08b9ec805bc766419d410748e2bea95846