Ensure the undercloud is deployed with the right mtu for ovb.
If performing an overcloud deploy with ovb the tasks in charge of
setting the right MTU in the undercloud would fail as they are
prepared for non-containerized undercloud, prior to OSP14. As from
the ovb official documentation [0] we can edit the mtu when deploying
the undercloud by configuring the local_mtu option in undercloud.conf.
Therefore, a task was added to ensure that the undercloud was
deployed with the right mtu, otherwise we'll fail.

[0] https://openstack-virtual-baremetal.readthedocs.io/en/latest/usage/usage.html

Change-Id: I814ee0ed42931dc567eb888fa8bfb4460b7a79e9