Fix OVB PXE boot
The instackenv-bmc.json.j2 conflated two concepts together -- IPMI
network and PXE boot network. The IPMI network is for power control
and we use that IPMI IP address for that. But the MAC address in
instackenv.json is for Ironic to identify the node when PXE booting,
so it must be on the network where Ironic listens, which is the
ctlplane network. The problem is now fixed, we use IPMI IP and PXE
MAC.

Another related change is fixing of order of NICs in the nodes
topology files. There should be no reason to use different NIC
ordering on different machines, so we converge on the order of
ovb_undercloud.

Co-Authored-By: Lukas Bezdicka <lbezdick@redhat.com>
Change-Id: I103e9d7ae3cceddd7438d05ac9fa583733dc4d2a